add (optional) src-refs
In some cases, it is a valueable metadata to know which components stem
from which sources. Add an optional src-refs attribute to resources that
allows to express this relationship.

Add an optional `labels` attribute to src-refs, to allow for the
specification of additional metadata (e.g. a subset of actually relevant
sources).