add optional component-name-mapping to repository-ctx
There are, unfortunately, OCI-Registries that do not support arbitrary
image-name-paths (e.g. docker-hub, registry.alicloud). Also, there are
retrictions as to the maximum allowed name-length (currently known for
registry.alicloud).

Add a new attribute specifying how component-names be mapped to
oci-artifacts, with a hash-digest-variant as an alternative to using the
original component-name.